Update to 20140221.

+++ sys/arch/i386/include/mcontext.h
@@ -86,7 +86,7 @@ typedef struct {
 		} __fp_xmm_state;	/* x87 and xmm regs in fxsave format */
 		int	__fp_fpregs[128];
 	} __fp_reg_set;
-	long	__fp_pad[33];			/* Historic padding */
+	int 	__fp_pad[33];			/* Historic padding */
 } __fpregset_t;
 __CTASSERT(sizeof (__fpregset_t) == 512 + 33 * 4);
